Issue #4 and #54: register for and import GPX and KML.
A bit of a hack, but this more or less follows the iOS model.  An abstract Telemetry base class, with two subclasses (GPXReader and KMLReader) that can read GPX and KML, respectively.

The hacky bits come from the fact that MFBMain has to respond to the incoming import request, including requesting read permissions.  Also because autototals is baked into ActNewFlight, I had to (slightly) re-invent auto-fill of cross-country and auto-fill of totals.

This does NOT handle all of the many flavors of KML.  KML without timestamps is useless and ignored.  Even KML with timestamps requires speed to be derived.

Still need to move export into these classes (despite saying "reader")

Fixed a few things, including:
 - Using the passed loc for takeoff/landing, rather than the current loc
 - LocSample had no getLocation method to return a Location object.
 - MFBLocation now has a flag to suppress recording, so that we don't overwrite an in-progress flight's telemetry with imported telemetry.  (We attach the raw imported telemetry afterwards)