Issue #1: outbox model.
Save/add inside ActNewFlight behave as follows:
 * Save for later -> QUEUED (always)
 * Add, Online -> Saves immediately, from the ActNewFlight screen
 * Add, Offline -> PENDING

When ActRecentsWS resumes, it tries to submit any PENDING flights, but not QUEUED flights.  If a flight fails to save, it is moved to QUEUED state (to avoid infinite loop) and error is persisted.

Persisting the error required changes to LogbookEntry and a schema change to the database, so bumped the version on that.  Since I needed to modify MFBConstants, I cleaned it up (slightly) too.

Fixed a bug as well where I wasn't showing any pending flights while offline.