Migrate away from deprecated StartActivityForResult and checkSelfPermission in mfbform, using the new architecture for pictures - ensures that the correct permissions are granted.
Calls "chooseImageCompleted", "takePictureCompleted", or "takeVideoCompleted" as appropriate; this is a no-op in ActMFBForm, needs to be subclassed for it to have any effect.

Also moved "AddPending further down in the menu list, since it was otherwsie awfully prominent and easy to hit by mistake"

Eliminated fFakePix, since I no longer use that.